fix: - split of carousel from main file by making a carousel from it
